add get_authorized_user helper function
